requirements.txt
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
pyowm==3.2.0
python-dateutil==2.8.2
python-editor==1.0.4
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
pandas==1.4.4
firebase-admin==5.0.0
twilio==7.8.0
requests==2.28.1
beautifulsoup4==4.9.3
lxml==4.9.0
Werkzeug==2.0.1
Pillow==9.5.0
opencv-python==4.8.0.76
scikit-learn==1.0.2
pyrebase4==4.7.1
face-recognition-models==0.3.0
face-recognition==1.3.0
dlib==19.24.0
